owner reservation: edit+delete security

+++ src/Krosbook/Controllers/Api/v1/CarReservationController.cs
@@ -90,6 +90,13 @@ namespace Krosbook.Controllers.Api.v1
         [ValidateModelState, CheckArgumentsForNull]    
         public IActionResult UpdateReservation(int reservationId, [FromBody] CarReservationViewModel reservationVm)
         {
+            if (reservationVm.UserId!=GetUserId()) {
+                var message = $"User with id "+GetUserId()+" can't update reservation, that was created by user with id "+reservationVm.UserId;
+                _logger.LogWarning(message);
+                this.Response.StatusCode = (int)HttpStatusCode.Unauthorized;
+                return this.Json(new { Message = message });
+            }
+
             if (reservationVm.Id != reservationId)
             {
                 var message = $"Invalid argument. Id '{reservationId}' and userVm.Id '{reservationVm.Id}' are not equal.";
@@ -119,6 +126,13 @@ namespace Krosbook.Controllers.Api.v1
         [HttpDelete("{reservationId}")]
         public IActionResult DeleteReservation(int reservationId)
         {
+            if (_reservationRepository.GetItem(reservationId).UserId != GetUserId())
+            {
+                var message = $"User with id " + GetUserId() + " can't delete reservation, that was created by user with id " + _reservationRepository.GetItem(reservationId).UserId;
+                _logger.LogWarning(message);
+                this.Response.StatusCode = (int)HttpStatusCode.Unauthorized;
+                return this.Json(new { Message = message });
+            }
             return SaveData(() =>
             {
                 _reservationRepository.Delete(reservationId);
